+++ Nightwatch.Samples.CSharp/SampleResource.cs
@@ -5,16 +5,19 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using JetBrains.Annotations;
+using Microsoft.FSharp.Core;
 using static Nightwatch.Core.Resources;
 
-namespace Nightwatch.Samples.CSharp
-{
-    public class SampleResource
-    {
-        public static ResourceFactory Factory { get; } =
-            new ResourceFactory("sample", CreateChecker);
+namespace Nightwatch.Samples.CSharp;
 
-        private static Func<Task<bool>> CreateChecker(IDictionary<string, string> param) =>
-            () => Task.FromResult(param["a"] == param["b"]);
-    }
+[PublicAPI]
+public class SampleResource
+{
+    public static ResourceFactory Factory { get; } = new("sample", CreateChecker);
+
+    private static Func<Task<FSharpResult<Unit, string>>> CreateChecker(IDictionary<string, string> param) =>
+        () => Task.FromResult(param["a"] == param["b"]
+            ? FSharpResult<Unit, string>.NewOk(null)
+            : FSharpResult<Unit, string>.NewError("Invalid state"));
 }